+++ demo-projects/meetup/site/pages/_app.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import App, { Container } from 'next/app';
+import App from 'next/app';
 import Head from 'next/head';
 import gql from 'graphql-tag';
 import { ApolloProvider } from 'react-apollo';
 import { ToastProvider } from 'react-toast-notifications';
 
 import withApollo from '../lib/withApollo';
-import { AuthProvider } from '../lib/authetication';
+import { AuthProvider } from '../lib/authentication';
 import StylesBase from '../primitives/StylesBase';
 import GoogleAnalytics from '../components/GoogleAnalytics';
 
@@ -38,21 +38,19 @@ class MyApp extends App {
     const { Component, pageProps, apolloClient, user } = this.props;
     return (
       <ToastProvider>
-        <Container>
-          <ApolloProvider client={apolloClient}>
-            <AuthProvider initialUserValue={user}>
-              <Head>
-                <link rel="shortcut icon" href="/favicon.ico" type="image/x-icon" />
-                <meta
-                  name="viewport"
-                  content="width=device-width, initial-scale=1, shrink-to-fit=no, viewport-fit=cover"
-                />
-              </Head>
-              <StylesBase />
-              <Component {...pageProps} />
-            </AuthProvider>
-          </ApolloProvider>
-        </Container>
+        <ApolloProvider client={apolloClient}>
+          <AuthProvider initialUserValue={user}>
+            <Head>
+              <link rel="shortcut icon" href="/favicon.ico" type="image/x-icon" />
+              <meta
+                name="viewport"
+                content="width=device-width, initial-scale=1, shrink-to-fit=no, viewport-fit=cover"
+              />
+            </Head>
+            <StylesBase />
+            <Component {...pageProps} />
+          </AuthProvider>
+        </ApolloProvider>
         <GoogleAnalytics />
       </ToastProvider>
     );